BREAKING CHANGE:
Changed the database scheme for comments. Instead of using a recursive
subdocument structure, every comment associated with a poem will hold a
reference to a parent comment.

If a comment is a top-level comment, the field repliesTo is set to null.

If a comment is a reply to another comment, the field repliesTo is set
to the string id of the parent comment.